fix: Enable webhooks when deploying with cert-manager
When using `make deploy` with the webhook/certmanager kustomize patches,
the operator was not starting the webhook server because ENABLE_WEBHOOKS
was not set in the deployment. This caused the E2E tests to fail with
"connection refused" errors when creating NotebookValidationJobs.

Add ENABLE_WEBHOOKS=true to the manager_webhook_patch.yaml so that
when webhooks are configured via kustomize, the operator actually
starts the webhook server.

Fixes: CrashLoopBackOff in E2E tests where webhook was configured
but not enabled in the operator binary.